killscrepepepppsaveconfig/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "discord-py>=2.5.2",
    "discord-py-self==2.0.1",
    "fastapi>=0.115.13",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "httpx>=0.28.1",
    "sqlalchemy>=2.0.41",
    "alembic>=1.16.2",
    "python-jose>=3.5.0",
    "passlib>=1.7.4",
]
